✨ Make interacting possible while drawer is opened Fixes #39
